chore: Update axiosInstance configuration and request interceptors
Refactor the axiosInstance configuration in axiosInstance.ts to include a request interceptor that adds the Authorization header with the token retrieved from the cookie. This ensures that authenticated requests are properly authorized. Also, update the API endpoints in useAnalyseResult.ts and useTakeAssessment.tsx to remove the leading slash to fix the request URLs.

Related to recent user commits:
- chore: Update npm dependency to latest version of @types/js-cookie
- bug fix
- Merge branch '7-convert-figma-to-code' into 15-integrate-ai-and-express-backend
- chore: Update useSignUp hook and server configuration